refactor: move render-iframe to commons
