+++ frontend/src/components/common/motd-modal/motd-modal.tsx
@@ -7,9 +7,9 @@ import { cypressId } from '../../../utils/cypress-attribute'
 import { Logger } from '../../../utils/logger'
 import { testId } from '../../../utils/test-id'
 import { EditorToRendererCommunicatorContextProvider } from '../../editor-page/render-context/editor-to-renderer-communicator-context-provider'
-import { RenderIframe } from '../../editor-page/renderer-pane/render-iframe'
 import { RendererType } from '../../render-page/window-post-message-communicator/rendering-message'
 import { CommonModal } from '../modals/common-modal'
+import { RenderIframe } from '../render-iframe/render-iframe'
 import { fetchMotd, MOTD_LOCAL_STORAGE_KEY } from './fetch-motd'
 import React, { useCallback, useMemo, useEffect, useState } from 'react'
 import { Button, Modal } from 'react-bootstrap'
