+++ frontend/src/components/common/render-iframe/hooks/use-send-additional-configuration-to-renderer.ts
