+++ frontend/src/components/common/render-iframe/render-iframe.tsx
@@ -7,8 +7,8 @@ import type { DarkModePreference } from '../../../redux/dark-mode/types'
 import { cypressAttribute, cypressId } from '../../../utils/cypress-attribute'
 import { Logger } from '../../../utils/logger'
 import { isTestMode } from '../../../utils/test-modes'
-import { ShowIf } from '../../common/show-if/show-if'
-import { WaitSpinner } from '../../common/wait-spinner/wait-spinner'
+import { useEditorToRendererCommunicator } from '../../editor-page/render-context/editor-to-renderer-communicator-context-provider'
+import type { ScrollProps } from '../../editor-page/synced-scroll/scroll-props'
 import { useExtensionEventEmitter } from '../../markdown-renderer/hooks/use-extension-event-emitter'
 import type { CommonMarkdownRendererProps } from '../../render-page/renderers/common-markdown-renderer-props'
 import { useEditorReceiveHandler } from '../../render-page/window-post-message-communicator/hooks/use-editor-receive-handler'
@@ -19,8 +19,8 @@ import type {
 } from '../../render-page/window-post-message-communicator/rendering-message'
 import type { RendererType } from '../../render-page/window-post-message-communicator/rendering-message'
 import { CommunicationMessageType } from '../../render-page/window-post-message-communicator/rendering-message'
-import { useEditorToRendererCommunicator } from '../render-context/editor-to-renderer-communicator-context-provider'
-import type { ScrollProps } from '../synced-scroll/scroll-props'
+import { ShowIf } from '../show-if/show-if'
+import { WaitSpinner } from '../wait-spinner/wait-spinner'
 import { useEffectOnRenderTypeChange } from './hooks/use-effect-on-render-type-change'
 import { useForceRenderPageUrlOnIframeLoadCallback } from './hooks/use-force-render-page-url-on-iframe-load-callback'
 import { useSendAdditionalConfigurationToRenderer } from './hooks/use-send-additional-configuration-to-renderer'
