+++ frontend/src/components/editor-page/editor-document-renderer/editor-document-renderer.tsx
@@ -6,9 +6,9 @@
 import { useApplicationState } from '../../../hooks/common/use-application-state'
 import { useTrimmedNoteMarkdownContentWithoutFrontmatter } from '../../../hooks/common/use-trimmed-note-markdown-content-without-frontmatter'
 import { setRendererStatus } from '../../../redux/renderer-status/methods'
+import type { RenderIframeProps } from '../../common/render-iframe/render-iframe'
+import { RenderIframe } from '../../common/render-iframe/render-iframe'
 import { RendererType } from '../../render-page/window-post-message-communicator/rendering-message'
-import type { RenderIframeProps } from '../renderer-pane/render-iframe'
-import { RenderIframe } from '../renderer-pane/render-iframe'
 import { useOnScrollWithLineOffset } from './hooks/use-on-scroll-with-line-offset'
 import { useScrollStateWithoutLineOffset } from './hooks/use-scroll-state-without-line-offset'
 import { NoteType } from '@hedgedoc/commons'
